third_party/blink/web_tests/platform/generic/touchadjustment/search-cancel-expected.txt

Test touch adjustment on a search field with a visible cancel button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Touch centered on cancel button but overlapping neighboring candidate.
PASS targetContainsPoint is true
PASS touchContainsPoint is true

Off-center touches enclosing cancel button.
PASS targetContainsPoint is true
PASS touchContainsPoint is true
PASS targetContainsPoint is true
PASS touchContainsPoint is true
PASS targetContainsPoint is true
PASS touchContainsPoint is true

Touches clipping left edge of cancel-button.
PASS targetContainsPoint is true
PASS touchContainsPoint is true
PASS targetContainsPoint is true
PASS touchContainsPoint is true
PASS targetContainsPoint is true
PASS touchContainsPoint is true
PASS successfullyParsed is true

TEST COMPLETE
